backend/requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
torch
torchvision
open_clip_torch
s2sphere
